store/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
packaging==21.3
pika==1.3.1
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.6
redis==4.3.5
sqlparse==0.4.3
PyJWT==2.6.0
